requirements.txt
numpy
Pillow
pickle-mixin
streamlit
scikit-learn==1.3.2
tensorflow==2.16.0rc0
